+++ lua/plugins/config/debugger/go.lua
@@ -1,5 +1,14 @@
 local dap = require('dap')
 
+local function dap_read(err, chunk)
+	assert(not err, err)
+	if chunk then
+		vim.schedule(function()
+			require('dap.repl').append(chunk)
+		end)
+	end
+end
+
 dap.adapters.go = function(callback, _config)
 	local stdout = vim.loop.new_pipe(false)
 	local stderr = vim.loop.new_pipe(false)
@@ -20,22 +29,8 @@ dap.adapters.go = function(callback, _config)
 		end
 	end)
 	assert(handle, 'Error running dlv: ' .. tostring(pid_or_err))
-	stdout:read_start(function(err, chunk)
-		assert(not err, err)
-		if chunk then
-			vim.schedule(function()
-				require('dap.repl').append(chunk)
-			end)
-		end
-	end)
-	stderr:read_start(function(err, chunk)
-		assert(not err, err)
-		if chunk then
-			vim.schedule(function()
-				require('dap.repl').append(chunk)
-			end)
-		end
-	end)
+	stdout:read_start(dap_read)
+	stderr:read_start(dap_read)
 	-- Wait for delve to start
 	vim.defer_fn(function()
 		callback({ type = 'server', host = '127.0.0.1', port = port })