+++ lua/plugins/config/focus_config.lua
@@ -16,6 +16,7 @@ return function(use)
 					'symboltree',
 					'filetree',
 				},
+				compatible_filetrees = { 'NvimTree' },
 			})
 		end,
 	})