+++ lua/plugins/config/marks_config.lua
@@ -1,0 +1,8 @@
+return function(use)
+	use({
+		'chentau/marks.nvim',
+		config = function()
+			require('marks').setup({})
+		end,
+	})
+end