+++ lua/settings/filetypes.lua
@@ -1,0 +1,1 @@
+vim.cmd([[au BufRead,BufNewFile *.mdpp set filetype=markdown]])