+++ lua/settings/init.lua
@@ -23,6 +23,7 @@ require('settings.search')
 require('settings.window')
 require('settings.neovide')
 require('settings.nvr')
+require('settings.filetypes')
 -- require('settings.backup')
 
 vim.cmd('set nobackup nowritebackup noswapfile noundofile')