+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Bug Fixes
 - **nvim-ts-rainbow:** froze to commit until problem is fixed
 
+### Code Refactoring
+- **debugger-go:** dryer code
+
 ### Features
 - **go:** test function now never cache
 - **golang:** change mappings for nvim goc