+++ CHANGELOG.md
@@ -11,12 +11,13 @@
 - **init:** refactor settings/init.lua
 
 ### Features
+- **filetype:** added certain filetype auto detect
 - **focus:** reenabled focus config
 - **go:** test function now never cache
 - **golang:** set golines limit to 200
 - **golang:** disabled golines and golangci-lint on nullls
-- **golang:** change mappings for nvim goc
 - **golang:** added debug current function
+- **golang:** change mappings for nvim goc
 - **marks:** update various marks file
 - **nvim:** max_line_lne set to 150
 - **nvim-tree:** update config